Posts GChild Frag Created and Added to Today Frag
The Posts fragment that will be a cardview of posts has been created. It
is being loaded up in the today fragment as a child fragment. The today
fragment's purpose will now be to retrieve the posts data for that
specific day, and simple pass that into the posts fragment to load up
the posts cards with those specific constraints in mind.