*BROKEN* PostComments Fragment added
added the postcomments fragment to show when a post is clicked. The Post
card will be shown at the top while there will be a second card, filled
with all the comments. NOTE: this is currently broken. The interface
being used to replace post fragment with the comments fragment is
raising an exception.