*FIXED* Comments Fragment now Opens
the post fragment gets replaced with a comments fragment when the
comment button is pressed. Currently, if you try to scroll, the app
crashed however this is simply because there is no adapter attached to
the comments recycler view at this moment. This issue will be fixed
soon. Transition animations also must be added.

NOTE: you need to PostFeedAdapter.setListener(this) and then implement
the postPressed method wherever you want to see the comments feed.