Fixed multitasking-crash bug & (almost) have backButtonPressed working
A bug that crashed the app when when you pressed the multitask button on
your device is now fixed. It required having several classes implement
Serializable and have some non-custom classes & instances of 3rd party
classes be transient. ALSO: the back button now successfully pops the
back stack in EACH TAB *EXCEPT PROFILE FRAG*. The downside to this is
that now each possible reusable frag must have its interfaces
implemented in each tab frag. However, this is likely to be the safest
and easiest method of doing so.